% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCog.R
\docType{data}
\name{dataCog}
\alias{dataCog}
\title{Made-up longitudinal data on global cognition before death}
\format{
A data frame with 11867 observations over 1200 subjects and 5 variables: \describe{
\item{ID}{subject identification number}
\item{time}{the retrospective (negative) time before death (in years)}
\item{cognition}{composite score of global cognition}
\item{ageDeath}{age at death (in years)}
\item{ageDeath90}{age at death centered by the mean}
}
}
\description{
The dataCog contains 1200 individuals with one line per visit. Repeated measures of global cognition (cognition) were collected over a maximum period of 30 years. Information on the age at death is also provided (ageDeath, ageDeath90).
}
\examples{

summary(dataCog)

}
\keyword{dataCog}
